Moved reddit tests to own module
